+++ packages/@magic-ext/farcaster/src/index.ts
@@ -1,132 +1,76 @@
-import { Extension, FarcasterLoginEventEmit, JsonRpcRequestPayload } from '@magic-sdk/commons';
-import type {
-  CreateChannelAPIResponse,
-  AuthenticateAPIResponse,
-  WatchStatusResponse,
-  AuthClientError,
-} from '@farcaster/auth-client';
-import { createAppClient, viemConnector } from '@farcaster/auth-client';
+import { Extension } from '@magic-sdk/commons';
+import type { CreateChannelAPIResponse, AuthenticateAPIResponse, AuthClientError } from '@farcaster/auth-client';
 import { FarcasterPayloadMethod } from './types';
 import { isMobile } from './utils';
 
-type Handle = {
-  on: Handler;
+const DEFAULT_SHOW_UI = true;
+
+type LoginParams = {
+  showUI: boolean;
 };
 
-const EVENT = {
-  CHANNEL: 'channel',
-  DONE: 'done',
-  ERROR: 'error',
+const FarcasterLoginEventOnReceived = {
+  OpenChannel: 'channel',
+  Success: 'success',
+  Failed: 'failed',
 } as const;
 
-type ChannelCallback = (params: CreateChannelAPIResponse) => void;
-type DoneCallback = (params: AuthenticateAPIResponse) => void;
-type ErrorCallback = (params: AuthClientError) => void;
+type FarcasterLoginEventHandlers = {
+  [FarcasterLoginEventOnReceived.OpenChannel]: (channel: CreateChannelAPIResponse) => void;
+  [FarcasterLoginEventOnReceived.Success]: (data: AuthenticateAPIResponse) => void;
+  [FarcasterLoginEventOnReceived.Failed]: (error: AuthClientError) => void;
+};
 
-const isChannelCallback = (callback: Function): callback is ChannelCallback => typeof callback === 'function';
-const isDoneCallback = (callback: Function): callback is DoneCallback => typeof callback === 'function';
-
-interface Handler {
-  (event: 'channel', callback: ChannelCallback): Handle;
-  (event: 'done', callback: DoneCallback): Handle;
-  (event: 'error', callback: ErrorCallback): Handle;
-}
-
-const DEFAULT_RELAY_URL = 'https://relay.farcaster.xyz';
-const DEFAULT_SIWE_URI = 'https://example.com/login';
-const DEFAULT_TIMEOUT = 60000;
-const DEFAULT_INTERVAL = 500;
-
+const FARCASTER_RELAY_URL = 'https://relay.farcaster.xyz';
 export class FarcasterExtension extends Extension.Internal<'farcaster'> {
   name = 'farcaster' as const;
   config = {};
+  channel: CreateChannelAPIResponse | null = null;
 
-  public login = ({ showUI }: { showUI: boolean }): Handle => {
-    const appClient = createAppClient({
-      relay: DEFAULT_RELAY_URL,
-      ethereum: viemConnector(),
-    });
+  constructor() {
+    super();
 
-    const channelPromise = appClient.createChannel({
-      siweUri: DEFAULT_SIWE_URI,
-      domain: window.location.host,
-    });
+    (async () => {
+      const json = await fetch(`${FARCASTER_RELAY_URL}/v1/channel`, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          domain: location.host,
+          siweUri: location.origin,
+        }),
+      }).then<CreateChannelAPIResponse>((r) => r.json());
 
-    const statusPromise: WatchStatusResponse = channelPromise.then(({ data }) => {
-      return appClient
-        .watchStatus({
-          channelToken: data.channelToken,
-          timeout: DEFAULT_TIMEOUT,
-          interval: DEFAULT_INTERVAL,
-        })
-        .then((r) => {
-          if (r.isError) {
-            throw r.error;
-          } else {
-            return r;
-          }
-        });
-    });
+      this.channel = json;
+    })();
+  }
 
-    let requestPayload: JsonRpcRequestPayload;
-    let channel_token: string;
+  public login = (params?: LoginParams) => {
+    if (!this.channel) {
+      throw new Error('Channel not created yet.');
+    }
 
-    const handle: Handle = {
-      on: (event, callback) => {
-        if (event === 'channel') {
-          (async () => {
-            const { data } = await channelPromise;
+    const showUI = params?.showUI ?? DEFAULT_SHOW_UI;
 
-            if (!isChannelCallback(callback)) return;
+    const domain = location.origin;
 
-            callback(data);
-
-            channel_token = data.channelToken;
-
-            if (isMobile()) {
-              console.info('Info: showUI parameter is ignored on mobile, open URL directly');
-              window.open(data.url, '_blank');
-              return;
-            }
-
-            requestPayload = this.utils.createJsonRpcRequestPayload(FarcasterPayloadMethod.FarcasterShowQR, [
-              { data: { showUI, ...data } },
-            ]);
-
-            this.request(requestPayload);
-          })();
-        }
-        if (event === EVENT.DONE) {
-          (async () => {
-            const { data } = await statusPromise;
-
-            if (data.state !== 'completed') return;
-
-            this.createIntermediaryEvent(
-              FarcasterLoginEventEmit.SuccessSignIn,
-              requestPayload.id as string,
-            )({
-              channel_token,
-              message: data.message,
-              signature: data.signature,
-              fid: data.fid,
-              username: data.username,
-            });
-
-            if (!isDoneCallback(callback)) return;
-
-            callback(data);
-          })();
-        }
-        if (event === EVENT.ERROR) {
-          statusPromise.catch((e) => {
-            callback(e);
-          });
-        }
-
-        return handle;
+    const payload = this.utils.createJsonRpcRequestPayload(FarcasterPayloadMethod.FarcasterShowQR, [
+      {
+        data: {
+          showUI,
+          domain,
+          isMobile: isMobile(),
+          channel: this.channel,
+        },
       },
-    };
+    ]);
+
+    const handle = this.request<string, FarcasterLoginEventHandlers>(payload);
+
+    if (isMobile()) {
+      location.href = this.channel.url;
+    }
 
     return handle;
   };