+++ packages/@magic-ext/farcaster/CHANGELOG.md
@@ -1,3 +1,62 @@
+# v0.0.6 (Fri Jun 28 2024)
+
+#### 🐛 Bug Fix
+
+- fix: fix mobile, move channel logic to magic-ext/farcaster [#755](https://github.com/magiclabs/magic-js/pull/755) ([@ysm-dev](https://github.com/ysm-dev))
+- fix: deepsource ([@ysm-dev](https://github.com/ysm-dev))
+- fix: fix mobile, move channel logic to magic-ext/farcaster ([@ysm-dev](https://github.com/ysm-dev))
+
+#### Authors: 1
+
+- Chris Yang ([@ysm-dev](https://github.com/ysm-dev))
+
+---
+
+# v0.0.5 (Wed Jun 26 2024)
+
+#### 🐛 Bug Fix
+
+- fix: move logic to outside of handle function [#754](https://github.com/magiclabs/magic-js/pull/754) ([@ysm-dev](https://github.com/ysm-dev))
+- fix: mobile ([@ysm-dev](https://github.com/ysm-dev))
+- fix: type ([@ysm-dev](https://github.com/ysm-dev))
+- refactor: move watch logic to mandrake ([@ysm-dev](https://github.com/ysm-dev))
+- Add console ([@ysm-dev](https://github.com/ysm-dev))
+- fix: move logic to outside of handle function ([@ysm-dev](https://github.com/ysm-dev))
+
+#### Authors: 1
+
+- Chris Yang ([@ysm-dev](https://github.com/ysm-dev))
+
+---
+
+# v0.0.4 (Mon Jun 24 2024)
+
+#### 🐛 Bug Fix
+
+- fix: iOS async popup [#753](https://github.com/magiclabs/magic-js/pull/753) ([@ysm-dev](https://github.com/ysm-dev))
+- fix: deepsource ([@ysm-dev](https://github.com/ysm-dev))
+- fix: rpc resolve timing ([@ysm-dev](https://github.com/ysm-dev))
+- fix: iOS async popup ([@ysm-dev](https://github.com/ysm-dev))
+
+#### Authors: 1
+
+- Chris Yang ([@ysm-dev](https://github.com/ysm-dev))
+
+---
+
+# v0.0.3 (Fri Jun 21 2024)
+
+#### 🐛 Bug Fix
+
+- Fix: farcaster mobile login [#752](https://github.com/magiclabs/magic-js/pull/752) ([@ysm-dev](https://github.com/ysm-dev))
+- Fix: farcaster mobile login ([@ysm-dev](https://github.com/ysm-dev))
+
+#### Authors: 1
+
+- Chris Yang ([@ysm-dev](https://github.com/ysm-dev))
+
+---
+
 # v0.0.2 (Thu Jun 20 2024)
 
 #### 🐛 Bug Fix